+++ modules/indexing/src/test/java/org/apache/ignite/internal/processors/query/timeout/DefaultQueryTimeoutThinJavaTest.java
@@ -23,6 +23,7 @@ import org.apache.ignite.cache.query.SqlFieldsQuery;
 import org.apache.ignite.client.IgniteClient;
 import org.apache.ignite.configuration.ClientConfiguration;
 import org.apache.ignite.internal.client.thin.ClientServerError;
+import org.apache.ignite.internal.processors.query.timeout.AbstractDefaultQueryTimeoutTest;
 import org.apache.ignite.internal.util.typedef.G;
 import org.apache.ignite.internal.util.typedef.X;
 
@@ -38,9 +39,9 @@ public class DefaultQueryTimeoutThinJavaTest extends AbstractDefaultQueryTimeout
     }
 
     /** {@inheritDoc} */
-    @Override protected void executeQuery(String sql, long timeout) throws Exception {
+    @Override protected void executeQuery(String sql, int timeout) throws Exception {
         try (IgniteClient cli = G.startClient(new ClientConfiguration().setAddresses("127.0.0.1"))) {
-            cli.query(new SqlFieldsQuery(sql).setTimeout((int)timeout, TimeUnit.MILLISECONDS)).getAll();
+            cli.query(new SqlFieldsQuery(sql).setTimeout(timeout, TimeUnit.MILLISECONDS)).getAll();
         }
     }
